Reorg some dataset stuff
