+++ src/opencosmo/collection/protocols.py
@@ -1,7 +1,7 @@
 from typing import Iterable, Protocol, Self, Union
 
 from opencosmo.dataset import Dataset
-from opencosmo.dataset.col import Mask
+from opencosmo.dataset.column import ColumnMask
 from opencosmo.header import OpenCosmoHeader
 from opencosmo.io.io import OpenTarget
 from opencosmo.io.protocols import DataSchema
@@ -44,7 +44,7 @@ class Collection(Protocol):
     def items(self) -> Iterable[tuple[str, Union[Dataset, "Collection"]]]: ...
     def __enter__(self): ...
     def __exit__(self, *exc_details): ...
-    def filter(self, *masks: Mask) -> Self: ...
+    def filter(self, *masks: ColumnMask) -> Self: ...
     def select(self, *args, **kwargs) -> Self: ...
     def with_units(self, convention: str) -> Self: ...
     def take(self, *args, **kwargs) -> Self: ...
