+++ src/opencosmo/dataset/__init__.py
@@ -1,4 +1,4 @@
-from .col import col
+from .column import col
 from .dataset import Dataset
 
 __all__ = ["col", "Dataset"]
