+++ src/opencosmo/dataset/handler.py
@@ -6,7 +6,7 @@ import h5py
 import numpy as np
 from astropy.table import Table  # type: ignore
 
-from opencosmo.dataset.column import TableBuilder
+from opencosmo.dataset.builders import TableBuilder
 from opencosmo.header import OpenCosmoHeader
 from opencosmo.index import DataIndex
 from opencosmo.io.schemas import DatasetSchema
